camera: basic implementation of camera

+++ src/buffer.cpp
@@ -3,7 +3,7 @@
 namespace luma {
 namespace buffer {
 
-vertex::vertex(void const* vertices, size_t const& size) {
+vertex::vertex(void const* vertices, uint32_t const& size) {
     glGenBuffers(1, &m_id);
     glBindBuffer(GL_ARRAY_BUFFER, m_id);
     glBufferData(GL_ARRAY_BUFFER, size, vertices, GL_STATIC_DRAW);
@@ -20,7 +20,7 @@ array::~array() {
     glDeleteVertexArrays(1, &m_id);
 }
 
-index::index(uint32_t const* indices, size_t const& count) : m_count(count) {
+index::index(uint32_t const* indices, uint32_t const& count) : m_count(count) {
     glGenBuffers(1, &m_id);
     glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, m_id);
     glBufferData(GL_ELEMENT_ARRAY_BUFFER, count * sizeof(uint32_t), indices, GL_STATIC_DRAW);
